Move serialization classes to common module
